src/requirements.txt
django
uvicorn
psycopg
pyjwt
djangorestframework